requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.15.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
gcsfs==2023.4.0
google-cloud-storage==2.9.0
greenlet==2.0.2
importlib-metadata==6.4.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
redis==4.5.4
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
